create Hotel.h
